grpc-server-backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangogrpcframework==0.2.1
djangorestframework==3.14.0
grpc-interceptor==0.15.4
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.3
psycopg2==2.9.10
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.1
